+++ scarletio/utils/multi_value_dictionary.py
@@ -117,7 +117,8 @@ class MultiValueDictionary(dict):
                 except KeyError:
                     setitem(self, key, [value])
                 else:
-                    values.append(value)
+                    if value not in values:
+                        values.append(value)
     
     
     @has_docs